new_steel_27_12/requirements.txt
Django==4.1
djangorestframework==3.13.1
drf-yasg==1.21.3
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
django-cors-headers==3.13.0
ipywidgets==7.6.5
scipy==1.9.1
sympy==1.11.1
Faker==14.2.0
whitenoise==6.2.0
environs==9.5.0
gunicorn==20.1.0
tabulate==0.8.10
psycopg2-binary==2.9.3
